Big bump
Now that the seg7multiplex project is working as intended, I can dust
off projects using it.

It's not quite working yet, but there's a lot of dusting off done,
namely upgrade to the C version of icemu. Also, it's the first time that
I integrate two circuits in the same simulation with the C version of
icemu. It's working quite well!